Change id to long && Enable block before QA Manager, EPG Leader, Conf. Manager

+++ achieveit_project/src/main/java/com/april/achieveit_project/entity/ActivityType.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ActivityType {
-    private Integer activityTypeId;
+    private Long activityTypeId;
 
     private String level1Description;
