+++ achieveit_userinfo/src/main/java/com/april/achieveit_userinfo/mapper/ProjectRoleMapper.java
@@ -3,13 +3,13 @@ package com.april.achieveit_userinfo.mapper;
 import com.april.achieveit_userinfo_interface.entity.ProjectRole;
 
 public interface ProjectRoleMapper {
-    int deleteByPrimaryKey(Integer projectRoleId);
+    int deleteByPrimaryKey(Long projectRoleId);
 
     int insert(ProjectRole record);
 
     int insertSelective(ProjectRole record);
 
-    ProjectRole selectByPrimaryKey(Integer projectRoleId);
+    ProjectRole selectByPrimaryKey(Long projectRoleId);
 
     int updateByPrimaryKeySelective(ProjectRole record);
