+++ achieveit_userinfo_interface/src/main/java/com/april/achieveit_userinfo_interface/entity/GlobalRole.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class GlobalRole
 {
-    private Integer globalRoleId;
+    private Long globalRoleId;
 
     private String globalRoleName;
 }
